+++ .github/workflows/linux.yml
@@ -12,20 +12,18 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - name: Install dependencies
-        run: sudo apt update && sudo apt install -y libssh2-1-dev libssl-dev
+        run: sudo apt update && sudo apt install -y libssh2-1-dev libssh-dev libssl-dev
       - uses: dtolnay/rust-toolchain@stable
         with:
           toolchain: stable
           components: rustfmt, clippy
       - name: Build
         run: cargo build --all-features
-      - name: Build (unsecure)
-        run: cargo build --no-default-features
       - name: Format
         run: cargo fmt --all -- --check
       - name: Clippy
-        run: cargo clippy --features find -- -Dwarnings
+        run: cargo lint
       - name: Run tests
-        run: cargo test --all-features --features find,github-actions
+        run: cargo test-all
         env:
           RUST_LOG: trace
