+++ Cargo.toml
@@ -10,7 +10,7 @@ license = "MIT"
 name = "remotefs-ssh"
 readme = "README.md"
 repository = "https://github.com/remotefs-rs/remotefs-rs-ssh"
-version = "0.6.4"
+version = "0.7.0"
 rust-version = "1.85.1"
 
 [dependencies]
@@ -19,7 +19,7 @@ lazy-regex = "3"
 log = "^0.4"
 remotefs = "^0.3"
 ssh2-config = "^0.6"
-ssh2 = "^0.9"
+ssh2 = { version = "^0.9", optional = true }
 
 [dev-dependencies]
 env_logger = "^0.11"
@@ -29,13 +29,11 @@ tempfile = "^3"
 testcontainers = { version = "0.25", features = ["blocking"] }
 
 [features]
-default = ["find"]
-# misc
+default = ["find", "libssh2"]
 find = ["remotefs/find"]
+libssh2 = ["dep:ssh2"]
+libssh2-vendored = ["libssh2", "ssh2/vendored-openssl"]
 no-log = ["log/max_level_off"]
-ssh2-vendored = ["ssh2/vendored-openssl"]
-# tests
-github-actions = []
 
 [target."cfg(target_os = \"windows\")"]
 [target."cfg(target_os = \"windows\")".dependencies]
