+++ README.md
@@ -11,7 +11,7 @@
 <p align="center">~ Remotefs SSH client ~</p>
 
 <p align="center">Developed by <a href="https://veeso.github.io/" target="_blank">@veeso</a></p>
-<p align="center">Current version: 0.6.4 (15/08/2025)</p>
+<p align="center">Current version: 0.7.0 (16/08/2025)</p>
 
 <p align="center">
   <a href="https://opensource.org/licenses/MIT"
@@ -72,14 +72,35 @@ First of all, add `remotefs-ssh` to your project dependencies:
 
 ```toml
 remotefs = "0.3"
-remotefs-ssh = "^0.6"
+remotefs-ssh = "^0.7"
 ```
+
+> ![NOTE]
+> The library supports multiple ssh backends.
+> Currently `libssh2` and `libssh` are supported.
+>
+> By default the library is using `libssh2`.
+
+### Available backends
+
+Each backend can be set as a feature in your `Cargo.toml`. Multiple backends can be enabled at the same time.
+
+- `libssh2`: The default backend, using the `libssh2` library for SSH connections.
+- `libssh`: An alternative backend, using the `libssh` library for SSH connections.
+
+Each backend can be built with the vendored version, using the vendored feature instead:
+
+- `libssh2-vendored`: Build the `libssh2` backend with the vendored version of the library.
+- `libssh-vendored`: Build the `libssh` backend with the vendored version of the library.
+
+If the vendored feature is **NOT** provided, you will need to have the corresponding system libraries installed on your machine.
+
+### Other features
 
 these features are supported:
 
 - `find`: enable `find()` method on client (*enabled by default*)
 - `no-log`: disable logging. By default, this library will log via the `log` crate.
-- `ssh2-vendored`: build with static libssl
 
 ---
