+++ src/mock.rs
@@ -14,6 +14,7 @@ pub fn logger() {
         let _ = env_logger::builder()
             .filter_level(log::LevelFilter::Trace)
             .is_test(true)
+            .format_source_path(true)
             .format_line_number(true)
             .try_init();
     });
