+++ src/ssh.rs
@@ -7,19 +7,20 @@ use std::path::{Path, PathBuf};
 use std::time::Duration;
 
 // -- modules
-mod commons;
+mod backend;
 mod config;
 #[cfg(test)]
 mod container;
+mod key_method;
 mod scp;
 mod sftp;
-mod stream;
-// -- export
-pub use scp::ScpFs;
-pub use sftp::SftpFs;
-pub use ssh2::MethodType as SshMethodType;
+
 pub use ssh2_config::ParseRule;
-use stream::{SftpReadStream, SftpWriteStream};
+
+pub use self::backend::SshSession;
+pub use self::key_method::{KeyMethod, MethodType};
+pub use self::scp::ScpFs;
+pub use self::sftp::SftpFs;
 
 // -- Ssh key storage
 
@@ -27,30 +28,6 @@ use stream::{SftpReadStream, SftpWriteStream};
 pub trait SshKeyStorage: Send + Sync {
     /// Return RSA key path from host and username
     fn resolve(&self, host: &str, username: &str) -> Option<PathBuf>;
-}
-
-// -- key method
-
-/// Ssh key method.
-/// Defined by [`MethodType`] (see ssh2 docs) and the list of supported algorithms.
-pub struct KeyMethod {
-    pub(crate) method_type: MethodType,
-    algos: Vec<String>,
-}
-
-impl KeyMethod {
-    /// Instantiates a new [`KeyMethod`]
-    pub fn new(method_type: MethodType, algos: &[String]) -> Self {
-        Self {
-            method_type,
-            algos: algos.to_vec(),
-        }
-    }
-
-    /// Get preferred algos in ssh protocol syntax
-    pub(crate) fn prefs(&self) -> String {
-        self.algos.join(",")
-    }
 }
 
 // -- ssh options
@@ -216,42 +193,6 @@ impl SshOpts {
     }
 }
 
-impl From<SshOpts> for SftpFs {
-    fn from(opts: SshOpts) -> Self {
-        SftpFs::new(opts)
-    }
-}
-
-impl From<SshOpts> for ScpFs {
-    fn from(opts: SshOpts) -> Self {
-        ScpFs::new(opts)
-    }
-}
-
-/// Re-implementation of ssh key method, in order to use `Eq`
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub enum MethodType {
-    CryptClientServer,
-    CryptServerClient,
-    HostKey,
-    Kex,
-    MacClientServer,
-    MacServerClient,
-}
-
-impl From<MethodType> for SshMethodType {
-    fn from(t: MethodType) -> Self {
-        match t {
-            MethodType::CryptClientServer => SshMethodType::CryptCs,
-            MethodType::CryptServerClient => SshMethodType::CryptSc,
-            MethodType::HostKey => SshMethodType::HostKey,
-            MethodType::Kex => SshMethodType::Kex,
-            MethodType::MacClientServer => SshMethodType::MacCs,
-            MethodType::MacServerClient => SshMethodType::MacSc,
-        }
-    }
-}
-
 #[cfg(test)]
 mod test {
 
@@ -331,15 +272,5 @@ mod test {
         );
         assert!(opts.key_storage.is_some());
         assert_eq!(opts.methods.len(), 1);
-    }
-
-    #[test]
-    fn should_build_sftp_client() {
-        let _: SftpFs = SshOpts::new("localhost").into();
-    }
-
-    #[test]
-    fn should_build_scp_client() {
-        let _: ScpFs = SshOpts::new("localhost").into();
     }
 }
