+++ src/ssh/scp.rs
@@ -2,7 +2,6 @@
 //!
 //! Scp remote fs implementation
 
-use std::io::{Read, Write};
 use std::ops::Range;
 use std::path::{Path, PathBuf};
 use std::time::{Duration, SystemTime};
@@ -13,10 +12,9 @@ use remotefs::fs::{
     FileType, Metadata, ReadStream, RemoteError, RemoteErrorType, RemoteFs, RemoteResult, UnixPex,
     UnixPexClass, Welcome, WriteStream,
 };
-// -- export
-pub use ssh2::Session as SshSession;
 
-use super::{SshOpts, commons};
+use super::SshOpts;
+use crate::SshSession;
 use crate::utils::{fmt as fmt_utils, parser as parser_utils, path as path_utils};
 
 /// NOTE: about this damn regex <https://stackoverflow.com/questions/32480890/is-there-a-regex-to-parse-the-values-from-an-ftp-directory-listing>
@@ -25,24 +23,33 @@ static LS_RE: Lazy<Regex> = lazy_regex!(
 );
 
 /// SCP "filesystem" client
-pub struct ScpFs {
-    session: Option<SshSession>,
+pub struct ScpFs<S>
+where
+    S: SshSession,
+{
+    session: Option<S>,
     wrkdir: PathBuf,
     opts: SshOpts,
 }
 
-impl ScpFs {
-    /// Creates a new `SftpFs`
-    pub fn new(opts: SshOpts) -> Self {
+#[cfg(feature = "libssh2")]
+impl ScpFs<super::backend::LibSsh2Session> {
+    /// Constructs a new [`ScpFs`] instance with the `libssh2` backend.
+    pub fn libssh2(opts: SshOpts) -> Self {
         Self {
             session: None,
             wrkdir: PathBuf::from("/"),
             opts,
         }
     }
+}
 
+impl<S> ScpFs<S>
+where
+    S: SshSession,
+{
     /// Get a reference to current `session` value.
-    pub fn session(&mut self) -> Option<&mut SshSession> {
+    pub fn session(&mut self) -> Option<&mut S> {
         self.session.as_mut()
     }
 
@@ -185,7 +192,7 @@ impl ScpFs {
 
     /// Execute setstat command and assert result is 0
     fn assert_stat_command(&mut self, cmd: String) -> RemoteResult<()> {
-        match commons::perform_shell_cmd_with_rc(self.session.as_mut().unwrap(), cmd) {
+        match self.session.as_mut().unwrap().cmd(cmd) {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::StatFailed)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -195,10 +202,12 @@ impl ScpFs {
     /// Returns whether file at `path` is a directory
     fn is_directory(&mut self, path: &Path) -> RemoteResult<bool> {
         let path = path_utils::absolutize(self.wrkdir.as_path(), path);
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("test -d \"{}\"", path.display()),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("test -d \"{}\"", path.display()))
+        {
             Ok((0, _)) => Ok(true),
             Ok(_) => Ok(false),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::StatFailed, err)),
@@ -206,20 +215,24 @@ impl ScpFs {
     }
 }
 
-impl RemoteFs for ScpFs {
+impl<S> RemoteFs for ScpFs<S>
+where
+    S: SshSession,
+{
     fn connect(&mut self) -> RemoteResult<Welcome> {
         debug!("Initializing SFTP connection...");
-        let mut session = commons::connect(&self.opts)?;
+        let mut session = S::connect(&self.opts)?;
         // Get banner
-        let banner: Option<String> = session.banner().map(String::from);
+        let banner = session.banner()?;
         debug!(
             "Connection established: {}",
             banner.as_deref().unwrap_or("")
         );
         // Get working directory
         debug!("Getting working directory...");
-        self.wrkdir = commons::perform_shell_cmd(&mut session, "pwd")
-            .map(|x| PathBuf::from(x.as_str().trim()))?;
+        self.wrkdir = session
+            .cmd("pwd")
+            .map(|(_rc, output)| PathBuf::from(output.as_str().trim()))?;
         // Set session
         self.session = Some(session);
         info!(
@@ -233,7 +246,7 @@ impl RemoteFs for ScpFs {
         debug!("Disconnecting from remote...");
         if let Some(session) = self.session.as_ref() {
             // Disconnect (greet server with 'Mandi' as they do in Friuli)
-            match session.disconnect(None, "Mandi!", None) {
+            match session.disconnect() {
                 Ok(_) => {
                     // Set session and sftp to none
                     self.session = None;
@@ -249,7 +262,7 @@ impl RemoteFs for ScpFs {
     fn is_connected(&mut self) -> bool {
         self.session
             .as_ref()
-            .map(|x| x.authenticated())
+            .map(|x| x.authenticated().unwrap_or_default())
             .unwrap_or(false)
     }
 
@@ -262,11 +275,19 @@ impl RemoteFs for ScpFs {
         self.check_connection()?;
         let dir = path_utils::absolutize(self.wrkdir.as_path(), dir);
         debug!("Changing working directory to {}", dir.display());
-        match commons::perform_shell_cmd(
-            self.session.as_mut().unwrap(),
-            format!("cd \"{}\"; echo $?; pwd", dir.display()),
-        ) {
-            Ok(output) => {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("cd \"{}\"; echo $?; pwd", dir.display()))
+        {
+            Ok((rc, output)) => {
+                if rc != 0 {
+                    return Err(RemoteError::new_ex(
+                        RemoteErrorType::ProtocolError,
+                        format!("Failed to change directory: {}", output),
+                    ));
+                }
                 // Trim
                 let output: String = String::from(output.as_str().trim());
                 // Check if output starts with 0; should be 0{PWD}
@@ -296,11 +317,19 @@ impl RemoteFs for ScpFs {
         if !self.exists(path.as_path()).ok().unwrap_or(false) {
             return Err(RemoteError::new(RemoteErrorType::NoSuchFileOrDirectory));
         }
-        match commons::perform_shell_cmd(
-            self.session.as_mut().unwrap(),
-            format!("unset LANG; ls -la \"{}/\"", path.display()).as_str(),
-        ) {
-            Ok(output) => {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("unset LANG; ls -la \"{}/\"", path.display()).as_str())
+        {
+            Ok((rc, output)) => {
+                if rc != 0 {
+                    return Err(RemoteError::new_ex(
+                        RemoteErrorType::ProtocolError,
+                        format!("Failed to list directory: {}", output),
+                    ));
+                }
                 // Split output by (\r)\n
                 let lines: Vec<&str> = output.as_str().lines().collect();
                 let mut entries: Vec<File> = Vec::with_capacity(lines.len());
@@ -331,8 +360,14 @@ impl RemoteFs for ScpFs {
             true => format!("ls -ld \"{}\"", path.display()),
             false => format!("ls -l \"{}\"", path.display()),
         };
-        match commons::perform_shell_cmd(self.session.as_mut().unwrap(), cmd.as_str()) {
-            Ok(line) => {
+        match self.session.as_mut().unwrap().cmd(cmd.as_str()) {
+            Ok((rc, line)) => {
+                if rc != 0 {
+                    return Err(RemoteError::new_ex(
+                        RemoteErrorType::NoSuchFileOrDirectory,
+                        format!("Failed to stat file: {line}"),
+                    ));
+                }
                 // Parse ls line
                 let parent: PathBuf = match path.as_path().parent() {
                     Some(p) => PathBuf::from(p),
@@ -355,10 +390,12 @@ impl RemoteFs for ScpFs {
     fn exists(&mut self, path: &Path) -> RemoteResult<bool> {
         self.check_connection()?;
         let path = path_utils::absolutize(self.wrkdir.as_path(), path);
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("test -e \"{}\"", path.display()),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("test -e \"{}\"", path.display()))
+        {
             Ok((0, _)) => Ok(true),
             Ok(_) => Ok(false),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::StatFailed, err)),
@@ -413,10 +450,12 @@ impl RemoteFs for ScpFs {
             return Err(RemoteError::new(RemoteErrorType::NoSuchFileOrDirectory));
         }
         debug!("Removing file {}", path.display());
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("rm -f \"{}\"", path.display()),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("rm -f \"{}\"", path.display()))
+        {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::CouldNotRemoveFile)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -430,10 +469,12 @@ impl RemoteFs for ScpFs {
             return Err(RemoteError::new(RemoteErrorType::NoSuchFileOrDirectory));
         }
         debug!("Removing directory {}", path.display());
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("rmdir \"{}\"", path.display()),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("rmdir \"{}\"", path.display()))
+        {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::DirectoryNotEmpty)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -447,10 +488,12 @@ impl RemoteFs for ScpFs {
             return Err(RemoteError::new(RemoteErrorType::NoSuchFileOrDirectory));
         }
         debug!("Removing directory {} recursively", path.display());
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("rm -rf \"{}\"", path.display()),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("rm -rf \"{}\"", path.display()))
+        {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::CouldNotRemoveFile)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -469,10 +512,11 @@ impl RemoteFs for ScpFs {
             path.display(),
             mode
         );
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("mkdir -m {} \"{}\"", mode, path.display()),
-        ) {
+        match self.session.as_mut().unwrap().cmd(format!(
+            "mkdir -m {} \"{}\"",
+            mode,
+            path.display()
+        )) {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::FileCreateDenied)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -493,10 +537,11 @@ impl RemoteFs for ScpFs {
         if self.exists(path.as_path()).ok().unwrap_or(false) {
             return Err(RemoteError::new(RemoteErrorType::FileCreateDenied));
         }
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("ln -s \"{}\" \"{}\"", target.display(), path.display()),
-        ) {
+        match self.session.as_mut().unwrap().cmd(format!(
+            "ln -s \"{}\" \"{}\"",
+            target.display(),
+            path.display()
+        )) {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new(RemoteErrorType::FileCreateDenied)),
             Err(err) => Err(RemoteError::new_ex(RemoteErrorType::ProtocolError, err)),
@@ -512,10 +557,12 @@ impl RemoteFs for ScpFs {
         }
         let dest = path_utils::absolutize(self.wrkdir.as_path(), dest);
         debug!("Copying {} to {}", src.display(), dest.display());
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("cp -rf \"{}\" \"{}\"", src.display(), dest.display()).as_str(),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("cp -rf \"{}\" \"{}\"", src.display(), dest.display()).as_str())
+        {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new_ex(
                 // Could not copy file
@@ -538,10 +585,12 @@ impl RemoteFs for ScpFs {
         }
         let dest = path_utils::absolutize(self.wrkdir.as_path(), dest);
         debug!("Moving {} to {}", src.display(), dest.display());
-        match commons::perform_shell_cmd_with_rc(
-            self.session.as_mut().unwrap(),
-            format!("mv -f \"{}\" \"{}\"", src.display(), dest.display()).as_str(),
-        ) {
+        match self
+            .session
+            .as_mut()
+            .unwrap()
+            .cmd(format!("mv -f \"{}\" \"{}\"", src.display(), dest.display()).as_str())
+        {
             Ok((0, _)) => Ok(()),
             Ok(_) => Err(RemoteError::new_ex(
                 // Could not copy file
@@ -555,11 +604,10 @@ impl RemoteFs for ScpFs {
     fn exec(&mut self, cmd: &str) -> RemoteResult<(u32, String)> {
         self.check_connection()?;
         debug!(r#"Executing command "{cmd}""#);
-        commons::perform_shell_cmd_at_with_rc(
-            self.session.as_mut().unwrap(),
-            cmd,
-            self.wrkdir.as_path(),
-        )
+        self.session
+            .as_mut()
+            .unwrap()
+            .cmd_at(cmd, self.wrkdir.as_path())
     }
 
     fn append(&mut self, _path: &Path, _metadata: &Metadata) -> RemoteResult<WriteStream> {
@@ -570,8 +618,6 @@ impl RemoteFs for ScpFs {
         self.check_connection()?;
         let path = path_utils::absolutize(self.wrkdir.as_path(), path);
         debug!("Creating file {}", path.display());
-        // blocking channel
-        self.session.as_mut().unwrap().set_blocking(true);
         trace!("blocked channel");
         let mode = metadata.mode.map(u32::from).unwrap_or(0o644) as i32;
         let accessed = metadata
@@ -595,7 +641,7 @@ impl RemoteFs for ScpFs {
             metadata.size,
             Some((modified, accessed)),
         ) {
-            Ok(channel) => Ok(WriteStream::from(Box::new(channel) as Box<dyn Write + Send>)),
+            Ok(channel) => Ok(WriteStream::from(channel)),
             Err(err) => {
                 error!("Failed to create file: {err}");
                 Err(RemoteError::new_ex(RemoteErrorType::FileCreateDenied, err))
@@ -611,10 +657,9 @@ impl RemoteFs for ScpFs {
         if !self.exists(path.as_path()).ok().unwrap_or(false) {
             return Err(RemoteError::new(RemoteErrorType::NoSuchFileOrDirectory));
         }
-        self.session.as_mut().unwrap().set_blocking(true);
         trace!("blocked channel");
         match self.session.as_mut().unwrap().scp_recv(path.as_path()) {
-            Ok((channel, _)) => Ok(ReadStream::from(Box::new(channel) as Box<dyn Read + Send>)),
+            Ok(channel) => Ok(ReadStream::from(channel)),
             Err(err) => {
                 error!("Failed to open file: {err}");
                 Err(RemoteError::new_ex(RemoteErrorType::CouldNotOpenFile, err))
@@ -624,976 +669,4 @@ impl RemoteFs for ScpFs {
 }
 
 #[cfg(test)]
-mod test {
-
-    use std::io::Cursor;
-
-    use pretty_assertions::assert_eq;
-    use ssh2_config::ParseRule;
-
-    use super::*;
-    use crate::mock::ssh as ssh_mock;
-    use crate::ssh::container::OpensshServer;
-
-    #[test]
-    fn should_init_scp_fs() {
-        let mut client = ScpFs::new(SshOpts::new("localhost"));
-        assert!(client.session.is_none());
-        assert_eq!(client.is_connected(), false);
-    }
-
-    #[test]
-    fn should_fail_connection_to_bad_server() {
-        let mut client = ScpFs::new(SshOpts::new("mybad.verybad.server"));
-        assert!(client.connect().is_err());
-    }
-
-    #[test]
-    fn should_not_append_to_file() {
-        crate::mock::logger();
-
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        // Append to file
-        let file_data = "Hello, world!\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        assert!(
-            client
-                .append_file(p, &Metadata::default(), Box::new(reader))
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_change_directory() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        let pwd = client.pwd().ok().unwrap();
-        assert!(client.change_dir(Path::new("/tmp")).is_ok());
-        assert!(client.change_dir(pwd.as_path()).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_change_directory_relative() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        assert!(
-            client
-                .create_dir(
-                    Path::new("should_change_directory_relative"),
-                    UnixPex::from(0o755)
-                )
-                .is_ok()
-        );
-        assert!(
-            client
-                .change_dir(Path::new("should_change_directory_relative/"))
-                .is_ok()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_change_directory() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        assert!(
-            client
-                .change_dir(Path::new("/tmp/sdfghjuireghiuergh/useghiyuwegh"))
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_copy_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        assert!(client.copy(p, Path::new("b.txt")).is_ok());
-        assert!(client.stat(p).is_ok());
-        assert!(client.stat(Path::new("b.txt")).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_copy_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        assert!(client.copy(p, Path::new("aaa/bbbb/ccc/b.txt")).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_create_directory() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // create directory
-        assert!(
-            client
-                .create_dir(Path::new("mydir"), UnixPex::from(0o755))
-                .is_ok()
-        );
-        let p = PathBuf::from(format!("{}/mydir", client.pwd().unwrap().display()));
-        assert!(client.exists(&p).unwrap());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_create_directory_cause_already_exists() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // create directory
-        assert!(
-            client
-                .create_dir(Path::new("mydir"), UnixPex::from(0o755))
-                .is_ok()
-        );
-        assert_eq!(
-            client
-                .create_dir(Path::new("mydir"), UnixPex::from(0o755))
-                .err()
-                .unwrap()
-                .kind,
-            RemoteErrorType::DirectoryAlreadyExists
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_create_directory() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // create directory
-        assert!(
-            client
-                .create_dir(
-                    Path::new("/tmp/werfgjwerughjwurih/iwerjghiwgui"),
-                    UnixPex::from(0o755)
-                )
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_create_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert_eq!(
-            client
-                .create_file(p, &metadata, Box::new(reader))
-                .ok()
-                .unwrap(),
-            10
-        );
-        // Verify size
-        assert_eq!(client.stat(p).ok().unwrap().metadata().size, 10);
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_create_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("/tmp/ahsufhauiefhuiashf/hfhfhfhf");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_exec_command() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        assert_eq!(
-            client.exec("echo 5").ok().unwrap(),
-            (0, String::from("5\n"))
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_tell_whether_file_exists() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        // Verify size
-        assert_eq!(client.exists(p).ok().unwrap(), true);
-        assert_eq!(client.exists(Path::new("b.txt")).ok().unwrap(), false);
-        assert_eq!(
-            client.exists(Path::new("/tmp/ppppp/bhhrhu")).ok().unwrap(),
-            false
-        );
-        assert_eq!(client.exists(Path::new("/tmp")).ok().unwrap(), true);
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_list_dir() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let wrkdir = client.pwd().ok().unwrap();
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        // Verify size
-        let file = client
-            .list_dir(wrkdir.as_path())
-            .ok()
-            .unwrap()
-            .first()
-            .unwrap()
-            .clone();
-        assert_eq!(file.name().as_str(), "a.txt");
-        let mut expected_path = wrkdir;
-        expected_path.push(p);
-        assert_eq!(file.path.as_path(), expected_path.as_path());
-        assert_eq!(file.extension().as_deref().unwrap(), "txt");
-        assert_eq!(file.metadata.size, 10);
-        assert_eq!(file.metadata.mode.unwrap(), UnixPex::from(0o644));
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_list_dir() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        assert!(client.list_dir(Path::new("/tmp/auhhfh/hfhjfhf/")).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_move_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        // Verify size
-        let dest = Path::new("b.txt");
-        assert!(client.mov(p, dest).is_ok());
-        assert_eq!(client.exists(p).ok().unwrap(), false);
-        assert_eq!(client.exists(dest).ok().unwrap(), true);
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_move_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        // Verify size
-        let dest = Path::new("/tmp/wuefhiwuerfh/whjhh/b.txt");
-        assert!(client.mov(p, dest).is_err());
-        assert!(
-            client
-                .mov(Path::new("/tmp/wuefhiwuerfh/whjhh/b.txt"), p)
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_open_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let metadata = Metadata::default().size(file_data.len() as u64);
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        // Verify size
-        let buffer: Box<dyn std::io::Write + Send> = Box::new(Vec::with_capacity(512));
-        assert_eq!(client.open_file(p, buffer).ok().unwrap(), 10);
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_open_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Verify size
-        let buffer: Box<dyn std::io::Write + Send> = Box::new(Vec::with_capacity(512));
-        assert!(
-            client
-                .open_file(Path::new("/tmp/aashafb/hhh"), buffer)
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_print_working_directory() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        assert!(client.pwd().is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_remove_dir_all() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create dir
-        let mut dir_path = client.pwd().ok().unwrap();
-        dir_path.push(Path::new("test/"));
-        assert!(
-            client
-                .create_dir(dir_path.as_path(), UnixPex::from(0o775))
-                .is_ok()
-        );
-        // Create file
-        let mut file_path = dir_path.clone();
-        file_path.push(Path::new("a.txt"));
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(
-            client
-                .create_file(file_path.as_path(), &metadata, Box::new(reader))
-                .is_ok()
-        );
-        // Remove dir
-        assert!(client.remove_dir_all(dir_path.as_path()).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_remove_dir_all() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Remove dir
-        assert!(
-            client
-                .remove_dir_all(Path::new("/tmp/aaaaaa/asuhi"))
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_remove_dir() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create dir
-        let mut dir_path = client.pwd().ok().unwrap();
-        dir_path.push(Path::new("test/"));
-        assert!(
-            client
-                .create_dir(dir_path.as_path(), UnixPex::from(0o775))
-                .is_ok()
-        );
-        assert!(client.remove_dir(dir_path.as_path()).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_remove_dir() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create dir
-        let mut dir_path = client.pwd().ok().unwrap();
-        dir_path.push(Path::new("test/"));
-        assert!(
-            client
-                .create_dir(dir_path.as_path(), UnixPex::from(0o775))
-                .is_ok()
-        );
-        // Create file
-        let mut file_path = dir_path.clone();
-        file_path.push(Path::new("a.txt"));
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(
-            client
-                .create_file(file_path.as_path(), &metadata, Box::new(reader))
-                .is_ok()
-        );
-        // Remove dir
-        assert!(client.remove_dir(dir_path.as_path()).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_remove_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.txt");
-        let file_data = "test data\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        assert!(client.remove_file(p).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_setstat_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.sh");
-        let file_data = "echo 5\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-
-        assert!(
-            client
-                .setstat(
-                    p,
-                    Metadata {
-                        accessed: Some(SystemTime::UNIX_EPOCH),
-                        created: None,
-                        file_type: FileType::File,
-                        gid: Some(1000),
-                        mode: Some(UnixPex::from(0o755)),
-                        modified: Some(SystemTime::UNIX_EPOCH),
-                        size: 7,
-                        symlink: None,
-                        uid: Some(1000),
-                    }
-                )
-                .is_ok()
-        );
-        let entry = client.stat(p).ok().unwrap();
-        let stat = entry.metadata();
-        assert_eq!(stat.accessed, None);
-        assert_eq!(stat.created, None);
-        assert_eq!(stat.modified, Some(SystemTime::UNIX_EPOCH));
-        assert_eq!(stat.mode.unwrap(), UnixPex::from(0o755));
-        assert_eq!(stat.size, 7);
-
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_setstat_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("bbbbb/cccc/a.sh");
-        assert!(
-            client
-                .setstat(
-                    p,
-                    Metadata {
-                        accessed: None,
-                        created: None,
-                        file_type: FileType::File,
-                        gid: Some(1),
-                        mode: Some(UnixPex::from(0o755)),
-                        modified: None,
-                        size: 7,
-                        symlink: None,
-                        uid: Some(1),
-                    }
-                )
-                .is_err()
-        );
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_stat_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.sh");
-        let file_data = "echo 5\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert_eq!(
-            client
-                .create_file(p, &metadata, Box::new(reader))
-                .ok()
-                .unwrap(),
-            7
-        );
-        let entry = client.stat(p).ok().unwrap();
-        assert_eq!(entry.name(), "a.sh");
-        let mut expected_path = client.pwd().ok().unwrap();
-        expected_path.push("a.sh");
-        assert_eq!(entry.path(), expected_path.as_path());
-        let meta = entry.metadata();
-        assert_eq!(meta.mode.unwrap(), UnixPex::from(0o644));
-        assert_eq!(meta.size, 7);
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_stat_file() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.sh");
-        assert!(client.stat(p).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_make_symlink() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.sh");
-        let file_data = "echo 5\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        let symlink = Path::new("b.sh");
-        assert!(client.symlink(symlink, p).is_ok());
-        assert!(client.remove_file(symlink).is_ok());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_not_make_symlink() {
-        crate::mock::logger();
-        let TestCtx {
-            mut client,
-            container: _container,
-        } = setup_client();
-        // Create file
-        let p = Path::new("a.sh");
-        let file_data = "echo 5\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        let mut metadata = Metadata::default();
-        metadata.size = file_data.len() as u64;
-        assert!(client.create_file(p, &metadata, Box::new(reader)).is_ok());
-        let symlink = Path::new("b.sh");
-        let file_data = "echo 5\n";
-        let reader = Cursor::new(file_data.as_bytes());
-        assert!(
-            client
-                .create_file(symlink, &metadata, Box::new(reader))
-                .is_ok()
-        );
-        assert!(client.symlink(symlink, p).is_err());
-        assert!(client.remove_file(symlink).is_ok());
-        assert!(client.symlink(symlink, Path::new("c.sh")).is_err());
-        finalize_client(client);
-    }
-
-    #[test]
-    fn should_get_name_and_link() {
-        let client = ScpFs::new(SshOpts::new("localhost"));
-        assert_eq!(
-            client.get_name_and_link("Cargo.toml"),
-            (String::from("Cargo.toml"), None)
-        );
-        assert_eq!(
-            client.get_name_and_link("Cargo -> Cargo.toml"),
-            (String::from("Cargo"), Some(PathBuf::from("Cargo.toml")))
-        );
-    }
-
-    #[test]
-    fn should_parse_file_ls_output() {
-        let client = ScpFs::new(SshOpts::new("localhost"));
-        // File
-        let entry = client
-            .parse_ls_output(
-                PathBuf::from("/tmp").as_path(),
-                "-rw-r--r-- 1 root root  2056 giu 13 21:11 /tmp/Cargo.toml",
-            )
-            .ok()
-            .unwrap();
-        assert_eq!(entry.name().as_str(), "Cargo.toml");
-        assert!(entry.is_file());
-        assert_eq!(entry.path, PathBuf::from("/tmp/Cargo.toml"));
-        assert_eq!(u32::from(entry.metadata.mode.unwrap()), 0o644_u32);
-        assert_eq!(entry.metadata.size, 2056);
-        assert_eq!(entry.extension().unwrap().as_str(), "toml");
-        assert!(entry.metadata.symlink.is_none());
-        // File (year)
-        let entry = client
-            .parse_ls_output(
-                PathBuf::from("/tmp").as_path(),
-                "-rw-rw-rw- 1 root root  3368 nov  7  2020 CODE_OF_CONDUCT.md",
-            )
-            .ok()
-            .unwrap();
-        assert_eq!(entry.name().as_str(), "CODE_OF_CONDUCT.md");
-        assert!(entry.is_file());
-        assert_eq!(entry.path, PathBuf::from("/tmp/CODE_OF_CONDUCT.md"));
-        assert_eq!(u32::from(entry.metadata.mode.unwrap()), 0o666_u32);
-        assert_eq!(entry.metadata.size, 3368);
-        assert_eq!(entry.extension().unwrap().as_str(), "md");
-        assert!(entry.metadata.symlink.is_none());
-    }
-
-    #[test]
-    fn should_parse_directory_from_ls_output() {
-        let client = ScpFs::new(SshOpts::new("localhost"));
-        // Directory
-        let entry = client
-            .parse_ls_output(
-                PathBuf::from("/tmp").as_path(),
-                "drwxr-xr-x 1 root root   512 giu 13 21:11 docs",
-            )
-            .ok()
-            .unwrap();
-        assert_eq!(entry.name().as_str(), "docs");
-        assert!(entry.is_dir());
-        assert_eq!(entry.path, PathBuf::from("/tmp/docs"));
-        assert_eq!(u32::from(entry.metadata.mode.unwrap()), 0o755_u32);
-        assert!(entry.metadata.symlink.is_none());
-        // Short metadata
-        assert!(
-            client
-                .parse_ls_output(
-                    PathBuf::from("/tmp").as_path(),
-                    "drwxr-xr-x 1 root root   512 giu 13 21:11",
-                )
-                .is_err()
-        );
-        // Special file
-        assert!(
-            client
-                .parse_ls_output(
-                    PathBuf::from("/tmp").as_path(),
-                    "crwxr-xr-x 1 root root   512 giu 13 21:11 ttyS1",
-                )
-                .is_err()
-        );
-        // Bad pex
-        assert!(
-            client
-                .parse_ls_output(
-                    PathBuf::from("/tmp").as_path(),
-                    "-rwxr-xr 1 root root   512 giu 13 21:11 ttyS1",
-                )
-                .is_err()
-        );
-    }
-
-    #[test]
-    fn should_parse_symlink_from_ls_output() {
-        let client = ScpFs::new(SshOpts::new("localhost"));
-        // File
-        let entry = client
-            .parse_ls_output(
-                PathBuf::from("/tmp").as_path(),
-                "lrw-r--r-- 1 root root  2056 giu 13 21:11 Cargo.toml -> Cargo.prod.toml",
-            )
-            .ok()
-            .unwrap();
-        assert_eq!(entry.name().as_str(), "Cargo.toml");
-        assert_eq!(entry.path, PathBuf::from("/tmp/Cargo.toml"));
-        assert_eq!(u32::from(entry.metadata.mode.unwrap()), 0o644_u32);
-        assert_eq!(entry.metadata.size, 2056);
-        assert_eq!(entry.extension().unwrap().as_str(), "toml");
-        assert_eq!(
-            entry.metadata.symlink.as_deref().unwrap(),
-            Path::new("Cargo.prod.toml")
-        );
-    }
-
-    #[test]
-    fn test_should_parse_special_permissions_ls_output() {
-        let client = ScpFs::new(SshOpts::new("localhost"));
-        assert!(
-            client
-                .parse_ls_output(
-                    Path::new("/tmp"),
-                    "-rw-rwSrw-    1 manufact  manufact    241813 Apr 22 09:31 L9800.SPF",
-                )
-                .is_ok()
-        );
-        assert!(
-            client
-                .parse_ls_output(
-                    Path::new("/tmp"),
-                    "-rw-rwsrw-    1 manufact  manufact    241813 Apr 22 09:31 L9800.SPF",
-                )
-                .is_ok()
-        );
-
-        assert!(
-            client
-                .parse_ls_output(
-                    Path::new("/tmp"),
-                    "-rw-rwtrw-    1 manufact  manufact    241813 Apr 22 09:31 L9800.SPF",
-                )
-                .is_ok()
-        );
-
-        assert!(
-            client
-                .parse_ls_output(
-                    Path::new("/tmp"),
-                    "-rw-rwTrw-    1 manufact  manufact    241813 Apr 22 09:31 L9800.SPF",
-                )
-                .is_ok()
-        );
-    }
-
-    #[test]
-    fn should_return_errors_on_uninitialized_client() {
-        let mut client = ScpFs::new(SshOpts::new("localhost"));
-        assert!(client.change_dir(Path::new("/tmp")).is_err());
-        assert!(
-            client
-                .copy(Path::new("/nowhere"), PathBuf::from("/culonia").as_path())
-                .is_err()
-        );
-        assert!(client.exec("echo 5").is_err());
-        assert!(client.disconnect().is_err());
-        assert!(client.list_dir(Path::new("/tmp")).is_err());
-        assert!(
-            client
-                .create_dir(Path::new("/tmp"), UnixPex::from(0o755))
-                .is_err()
-        );
-        assert!(client.symlink(Path::new("/a"), Path::new("/b")).is_err());
-        assert!(client.pwd().is_err());
-        assert!(client.remove_dir_all(Path::new("/nowhere")).is_err());
-        assert!(
-            client
-                .mov(Path::new("/nowhere"), Path::new("/culonia"))
-                .is_err()
-        );
-        assert!(client.stat(Path::new("/tmp")).is_err());
-        assert!(
-            client
-                .setstat(Path::new("/tmp"), Metadata::default())
-                .is_err()
-        );
-        assert!(client.open(Path::new("/tmp/pippo.txt")).is_err());
-        assert!(
-            client
-                .create(Path::new("/tmp/pippo.txt"), &Metadata::default())
-                .is_err()
-        );
-        assert!(
-            client
-                .append(Path::new("/tmp/pippo.txt"), &Metadata::default())
-                .is_err()
-        );
-    }
-
-    fn is_send<T: Send>(_send: T) {}
-
-    fn is_sync<T: Sync>(_sync: T) {}
-
-    #[test]
-    fn test_should_be_sync() {
-        let client = ScpFs::new(
-            SshOpts::new("sftp").key_storage(Box::new(ssh_mock::MockSshKeyStorage::default())),
-        );
-
-        is_sync(client);
-    }
-
-    #[test]
-    fn test_should_be_send() {
-        let client = ScpFs::new(
-            SshOpts::new("sftp").key_storage(Box::new(ssh_mock::MockSshKeyStorage::default())),
-        );
-        is_send(client);
-    }
-
-    // -- test utils
-
-    struct TestCtx {
-        client: ScpFs,
-        #[allow(dead_code)]
-        container: OpensshServer,
-    }
-
-    fn setup_client() -> TestCtx {
-        let container = OpensshServer::start();
-        let port = container.port();
-
-        use crate::SshAgentIdentity;
-
-        let config_file = ssh_mock::create_ssh_config(port);
-        let mut client = ScpFs::new(
-            SshOpts::new("scp")
-                .key_storage(Box::new(ssh_mock::MockSshKeyStorage::default()))
-                .config_file(config_file.path(), ParseRule::ALLOW_UNKNOWN_FIELDS)
-                .ssh_agent_identity(Some(SshAgentIdentity::All)),
-        );
-        assert!(client.connect().is_ok());
-        // Create wrkdir
-        let tempdir = PathBuf::from(generate_tempdir());
-        assert!(
-            client
-                .create_dir(tempdir.as_path(), UnixPex::from(0o775))
-                .is_ok()
-        );
-        // Change directory
-        assert!(client.change_dir(tempdir.as_path()).is_ok());
-
-        TestCtx { client, container }
-    }
-
-    fn finalize_client(mut client: ScpFs) {
-        // Get working directory
-        let wrkdir = client.pwd().ok().unwrap();
-        // Remove directory
-        assert!(client.remove_dir_all(wrkdir.as_path()).is_ok());
-        assert!(client.disconnect().is_ok());
-    }
-
-    fn generate_tempdir() -> String {
-        use rand::distr::Alphanumeric;
-        use rand::{Rng, rng};
-        let mut rng = rng();
-        let name: String = std::iter::repeat(())
-            .map(|()| rng.sample(Alphanumeric))
-            .map(char::from)
-            .take(8)
-            .collect();
-        format!("/tmp/temp_{name}")
-    }
-}
+mod tests;
