+++ src/ssh/scp/tests.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "libssh2")]
+mod libssh2;
+
+use super::super::backend::LibSsh2Session;
+use super::*;
