+++ src/ssh/sftp/tests.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "libssh2")]
+mod libssh2;
+
+use super::super::backend::*;
+use super::*;
